--- Assets/Events/DragEvents.cs
@@ -1,43 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-using UnityEngine.EventSystems;
-
-public class DragEvents : MonoBehaviour, IPointerDownHandler, IBeginDragHandler, IEndDragHandler, IDragHandler
-{
-    private RectTransform rectTransform;
-    private CanvasGroup canvasGroup;
-    private Vector3 position;
-    private void Awake()
-    {
-        rectTransform = GetComponent<RectTransform>();
-        canvasGroup = GetComponent<CanvasGroup>();
-        //TODO change smth about card apperaing
-        position = rectTransform.localPosition;
-        position.x -= 480;
-        position.y -= 15;
-    }
-    public void OnBeginDrag(PointerEventData eventData)
-    {
-        canvasGroup.alpha = .6f;
-        canvasGroup.blocksRaycasts = false;
-    }
-
-    public void OnEndDrag(PointerEventData eventData)
-    {
-        canvasGroup.alpha = 1f;
-        rectTransform.anchoredPosition = position;
-        canvasGroup.blocksRaycasts = true;
-        Debug.Log("OnEndDrag");
-    }
-
-    public void OnPointerDown(PointerEventData eventData)
-    {
-        Debug.Log("OnPointerDown");
-    }
-
-    public void OnDrag(PointerEventData eventData)
-    {
-        rectTransform.anchoredPosition += eventData.delta;
-    }
-}
