--- Assets/Events/DragEvents.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 2a65065be213cb84eb2cbcf47998e120
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
